Refactored to use a TokenService, moved token replacement to SearchService
