+++ solutions/ModernSearch/react-search-refiners/spfx/src/services/TokenService/ITokenService.ts
@@ -1,0 +1,3 @@
+export interface ITokenService {
+    replaceQueryVariables(queryTemplate: string): Promise<string>;
+}
