+++ solutions/ModernSearch/react-search-refiners/spfx/src/services/TokenService/index.ts
@@ -1,0 +1,2 @@
+export * from './ITokenService';
+export * from './TokenService';
